Integrate `@JsonTypeInfo` tests
