+++ test-suite-tck-jackson-databind/src/test/groovy/io/micronaut/serde/tck/jackson/databind/DatabindJsonTypeInfoSpec.groovy
@@ -1,8 +1,11 @@
 package io.micronaut.serde.tck.jackson.databind
 
 import io.micronaut.serde.jackson.JsonTypeInfoSpec
-import spock.lang.Ignore
 
-@Ignore // TODO
 class DatabindJsonTypeInfoSpec extends JsonTypeInfoSpec {
+
+    @Override
+    protected boolean jacksonCustomOrder() {
+        return true
+    }
 }
